services/python/requirements.txt
langextract
python-dotenv
